* Moving data-types defined for state manager back within the scope of the core package and out of utils. * Moving range-store data types into range-store interface to keep interface-level data-type access from referencing implementation code * Adding default StateManager with dummy implementation
