+++ packages/core/src/interfaces/client/index.ts
@@ -1,0 +1,1 @@
+export * from './state-manager.interface'
