+++ packages/core/src/interfaces/common/utils/range-store.interface.ts
@@ -1,0 +1,11 @@
+
+import BigNum = require('bn.js')
+
+export interface Range {
+  start: BigNum
+  end: BigNum
+}
+
+export interface BlockRange extends Range {
+  block: BigNum
+}
